Added a possible approach for balancing
This documents describes two possible approaches on how the blockchain would actually enable balancing. presumably we're not going to implement this as a core feature (unless we have a of a lot of time extra of course). But at least this shows that our blockchain would be technically capable of balancing, and how it should be implemented.